+++ src/lib.rs
@@ -53,6 +53,7 @@ mod env;
 mod value;
 mod types;
 mod error;
+mod call;
 
 /// This exposes some raw types for module to use (e.g. in `emacs_module_init`) without having to
 /// declare the raw `emacs_module` as a dependency.